pyproject.toml
[project]
name = "autoforge-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "langchain>=1.2.14",
    "langchain-openai>=1.1.12",
    "mcp[cli]>=1.26.0",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
